requirements.txt
streamlit
requests
beautifulsoup4
fpdf
numpy
scikit-learn
python-whois
streamlit-lottie
pandas
